requirements.txt
fastapi==0.115.6
pydantic==2.10.5
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.34.0
email-validator==2.2.0
fastapi-pagination==0.12.34
sqlmodel==0.0.22
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
Faker==35.0.0